More tests and lints

+++ Jax/Jax.php
@@ -91,6 +91,8 @@ final readonly class Jax
     }
 
     /**
+     * Given the total pages, active page, and number of links you want
+     * Returns an array that always has the first and last page, with the active page in the middle.
      * @return array<int>
      */
     public function pages(int $numpages, int $active, int $tofill): array
@@ -101,7 +103,7 @@ final readonly class Jax
             return $pages;
         }
 
-        $start = $active - floor($tofill / 2);
+        $start = $active - (int) floor($tofill / 2);
         if ($numpages - $start < $tofill) {
             $start -= $tofill - ($numpages - $start);
         }
